# services/gateway/openapi.yaml
openapi: "3.1.0"
info:
  title: KVYT API
  version: "1.0.0"
  description: |
    Public API of KVYT, an event ticketing system: browse events and sessions,
    hold seats, confirm or cancel bookings.

    All paths start with `/api/v1`. Every response carries an `X-Trace-Id` header;
    a client may send its own `X-Trace-Id` and the system will use it.

    Errors share one format (`Problem`): a stable machine-readable `code` and a
    human-readable `message`.

    Endpoints under `/me` and `/bookings` require an access token from
    `POST /api/v1/auth/login`, sent as `Authorization: Bearer <access_token>`.
    An access token lives 15 minutes; get a new pair with `POST /api/v1/auth/refresh`.

    If an internal service is temporarily unavailable, any endpoint may answer
    `502`, `503` or `504` with a `Problem` body.
servers:
  - url: /
...
paths:
...
  /api/v1/events:
    get:
      tags: [events]
      operationId: listEvents
      summary: List events
      description: |
        Events sorted by title. Paginated with `limit` and `offset`; `total` is the
        number of events matching the filters, regardless of the page.
      parameters:
        - name: city
          in: query
          description: City, case-insensitive exact match
          schema:
            type: string
            maxLength: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [published, archived]
            default: published
        - name: limit
          in: query
          description: Page size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of events to skip from the start of the list
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: One page of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPage"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/events/{event_id}:
    parameters:
      - $ref: "#/components/parameters/EventId"
    get:
      tags: [events]
      operationId: getEvent
      summary: Get an event
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventOut"
        "404":
          description: "No such public event (`event_not_found`)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/events/{event_id}/sessions:
    parameters:
      - $ref: "#/components/parameters/EventId"
    get:
      tags: [events]
      operationId: listEventSessions
      summary: Sessions of an event
      description: Sorted by start time.
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionOut"
        "404":
          description: "No such public event (`event_not_found`)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `POST /api/v1/auth/login`.
  parameters:
    EventId:
      name: event_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    ValidationError:
      description: "The request does not match the schema (`unprocessable_entity`)"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      description: Error body, the same for every endpoint.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id]
          properties:
            code:
              type: string
              description: Stable machine-readable code
            message:
              type: string
              description: Human-readable explanation
            trace_id:
              type: string
              description: Same as the `X-Trace-Id` response header
...
    EventOut:
      type: object
      required: [id, title, venue, city, description, organizer_id, status]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        venue:
          type: string
        city:
          type: string
        description:
          type: string
        organizer_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [published, archived]
    EventPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EventOut"
        total:
          type: integer
          minimum: 0
          description: Number of events matching the filters, on all pages
        limit:
          type: integer
          description: Page size applied by the server
        offset:
          type: integer
          description: Offset applied by the server
    SessionOut:
      type: object
      required: [id, event_id, starts_at, base_price_uah]
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        starts_at:
          type: string
          format: date-time
        base_price_uah:
          type: integer
          description: Price of the cheapest seat category, in hryvnias
...